rename types file names
