+++ src/pages/publication/submission_tab/submission_journal.tsx
@@ -3,7 +3,7 @@ import React, { FC } from 'react';
 import { useTranslation } from 'react-i18next';
 import { FormikProps, useFormikContext } from 'formik';
 import { Publication, PagesRange } from '../../../types/publication.types';
-import { JournalEntityDescription } from '../../../types/publication_types/journal.publication.types';
+import { JournalEntityDescription } from '../../../types/publication_types/journalPublication.types';
 
 const SubmissionJournalPublication: FC = () => {
   const { t } = useTranslation('publication');
