+++ src/pages/publication/submission_tab/submission_report.tsx
@@ -3,7 +3,7 @@ import React, { FC } from 'react';
 import { useTranslation } from 'react-i18next';
 import { FormikProps, useFormikContext } from 'formik';
 import { Publication } from '../../../types/publication.types';
-import { ReportEntityDescription } from '../../../types/publication_types/report.publication.types';
+import { ReportEntityDescription } from '../../../types/publication_types/reportPublication.types';
 
 const SubmissionReport: FC = () => {
   const { t } = useTranslation('publication');
