+++ src/types/publication_types/bookPublication.types.ts
@@ -2,7 +2,7 @@ import { BackendType, PublicationDate, PagesMonograph } from '../publication.typ
 import { PublicationType, BookType } from '../publicationFieldNames';
 import { LanguageValues } from '../language.types';
 import { Contributor } from '../contributor.types';
-import { BackendTypeNames, emptyDate } from './common.publication.types';
+import { BackendTypeNames, emptyDate } from './commonPublication.types';
 
 interface BookPublicationInstance {
   type: BookType | '';
