+++ src/types/publication_types/commonPublication.types.ts
