+++ src/types/publication_types/reportPublication.types.ts
@@ -2,32 +2,35 @@ import { BackendType, PublicationDate, PagesMonograph } from '../publication.typ
 import { PublicationType, DegreeType } from '../publicationFieldNames';
 import { LanguageValues } from '../language.types';
 import { Contributor } from '../contributor.types';
-import { BackendTypeNames, emptyDate } from './common.publication.types';
+import { BackendTypeNames, emptyDate } from './commonPublication.types';
 
-interface DegreePublicationInstance {
+interface ReportPublicationInstance {
   type: DegreeType | '';
   pages: PagesMonograph | null;
   peerReviewed: boolean;
 }
 
-export interface DegreePublicationContext {
+export interface ReportPublicationContext {
   type: PublicationType | '';
   isbnList: string[];
+  level: string | number | null;
+  onlineIssn: string;
   openAccess: boolean;
   peerReviewed: boolean;
+  printIssn: string;
   publisher: string;
   seriesNumber: string;
   seriesTitle: string;
   url: string;
 }
 
-interface DegreeReference extends BackendType {
+interface ReportReference extends BackendType {
   doi: string;
-  publicationContext: DegreePublicationContext;
-  publicationInstance: DegreePublicationInstance;
+  publicationContext: ReportPublicationContext;
+  publicationInstance: ReportPublicationInstance;
 }
 
-export interface DegreeEntityDescription extends BackendType {
+export interface ReportEntityDescription extends BackendType {
   abstract: string;
   contributors: Contributor[];
   date: PublicationDate;
@@ -35,35 +38,38 @@ export interface DegreeEntityDescription extends BackendType {
   language: LanguageValues;
   mainTitle: string;
   npiSubjectHeading: string;
-  reference: DegreeReference;
+  reference: ReportReference;
   tags: string[];
 }
 
-export const emptyPublicationInstance: DegreePublicationInstance = {
+export const emptyPublicationInstance: ReportPublicationInstance = {
   type: '',
   pages: null,
   peerReviewed: false,
 };
 
-export const emptyPublicationContext: DegreePublicationContext = {
+export const emptyPublicationContext: ReportPublicationContext = {
   type: '',
   isbnList: [],
+  level: '',
+  onlineIssn: '',
   openAccess: false,
   peerReviewed: false,
+  printIssn: '',
   publisher: '',
   seriesNumber: '',
   seriesTitle: '',
   url: '',
 };
 
-export const emptyReference: DegreeReference = {
+export const emptyReference: ReportReference = {
   type: BackendTypeNames.REFERENCE,
   doi: '',
   publicationContext: emptyPublicationContext,
   publicationInstance: emptyPublicationInstance,
 };
 
-export const emptyPublicationEntityDescription: DegreeEntityDescription = {
+export const emptyPublicationEntityDescription: ReportEntityDescription = {
   type: BackendTypeNames.ENTITY_DESCRIPTION,
   abstract: '',
   contributors: [],
